move the teambinding code to a separate dir
